Misc. adjustments to file uploads
