+++ .dockerignore
@@ -4,3 +4,5 @@ coverage/
 logs/
 *.log
 .vscode
+__tests__/
+__test__/
