+++ Dockerfile
@@ -16,10 +16,9 @@ COPY ./config/default.js ./config/default.js
 RUN npm install --only=prod
 
 COPY --from=builder /usr/src/app/dist .
-COPY --from=builder /usr/src/app/src/infrastructure/grpc/protos/*.proto ./infrastructure/grpc/protos/
 EXPOSE 8888
 
-CMD [ "node", "./server.js" ]
+CMD [ "node", "./presentation/server.js" ]
 # To ship logs to the ELK stack extend the above command
 # with either pino-socket, pino-logstash or mds-log-pump.
 # An example using mds-log-pump can be found in the mds
@@ -30,6 +29,6 @@ CMD [ "node", "./server.js" ]
 # If you chose to use pino-socket or pino-logstash you will
 # need to refer to their documentation for configuration.
 #
-# Ex: CMD [ "node", "./server.js", "|", "mds-log-pump"]
-# Ex: CMD [ "node", "./server.js", "|", "pino-logstash", "-h", "elk", "-p", "5000" ]
-# Ex: CMD [ "node", "./server.js", "|", "pino-socket", "-h", "elk", "-p", "5000" ]
+# Ex: CMD [ "node", "./presentation/server.js", "|", "mds-log-pump"]
+# Ex: CMD [ "node", "./presentation/server.js", "|", "pino-logstash", "-h", "elk", "-p", "5000" ]
+# Ex: CMD [ "node", "./presentation/server.js", "|", "pino-socket", "-h", "elk", "-p", "5000" ]
