--- bin/server
@@ -1,51 +1,0 @@
-#!/usr/bin/env node
-
-const _ = require('lodash');
-const mdsSdk = require('@maddonkeysoftware/mds-cloud-sdk-node');
-
-const src = require('../src/presentation');
-const globals = require('../src/globals');
-const runtimes = require('../src/runtimes');
-const containerManager = require('../src/containerManager');
-const shutdownHandler = require('../src/handlers/app_shutdown');
-const repo = require('../src/infrastructure/repos/mongo-repo-base');
-
-const logger = globals.getLogger();
-const specifiedRuntimes = _.get(process.env, ['MDS_FN_RUNTIMES'], '').split(',');
-
-const init = async () => {
-  if (runtimes.reduce(specifiedRuntimes).length === 0) {
-    logger.fatal({
-      validRuntimes: runtimes.SUPPORTED_RUNTIMES,
-      specifiedRuntimes,
-      environmentSetting: _.get(process.env, ['MDS_FN_RUNTIMES'], ''),
-    }, 'Missing valid value(s) for environment variable MDS_FN_RUNTIMES');
-  } else {
-    const conf = {
-      nsUrl: process.env.MDS_FN_NS_URL,
-      qsUrl: process.env.MDS_FN_QS_URL,
-      fsUrl: process.env.MDS_FN_FS_URL,
-      identityUrl: process.env.MDS_IDENTITY_URL,
-      account: process.env.MDS_FN_SYS_ACCOUNT,
-      userId: process.env.MDS_FN_SYS_USER,
-      password: process.env.MDS_FN_SYS_PASSWORD,
-    };
-    await mdsSdk.MdsSdk.initialize(conf);
-    logger.info({
-      config: _.omit(conf, 'password'),
-      specifiedRuntimes,
-    }, 'Application initialized');
-
-    containerManager.startMonitor();
-    shutdownHandler.wire(() => {
-      containerManager.stopMonitor();
-    });
-    const port = process.env.APP_PORT || 8888;
-    const app = src.buildApp();
-    repo.setupIndexes();
-
-    app.listen(port, () => logger.info(`MDS Cloud Docker Minion app listening on port ${port}!`));
-  }
-};
-
-init();
