+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "A worker minion to handle docker specific work from the mdsCloudServerlessFunctions application",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json",
+    "build": "tsc -p tsconfig.build.json && copyfiles -f -V ./src/infrastructure/grpc/protos/*.proto ./dist/infrastructure/grpc/protos",
     "clean": "rimraf ./dist",
     "dev": "concurrently -c green,blue -n server,db \"wait-on tcp:27017 && npm run dev:server\" \"docker compose down -v && docker compose up -V\"",
     "dev:server": "NODE_ENV=localdev ts-node-dev --respawn --pretty --transpile-only src/presentation/server.ts",
@@ -33,7 +33,7 @@
   "homepage": "https://github.com/MadDonkeySoftware/mdsCloudDockerMinion#readme",
   "dependencies": {
     "@fastify/awilix": "^3.1.0",
-    "@fastify/multipart": "^7.6.0",
+    "@fastify/multipart": "^7.7.3",
     "@fastify/swagger": "^8.7.0",
     "@fastify/swagger-ui": "^1.9.2",
     "@fastify/type-provider-typebox": "^2.4.0",
@@ -69,6 +69,7 @@
     "@typescript-eslint/eslint-plugin": "^5.59.9",
     "@typescript-eslint/parser": "^5.59.9",
     "concurrently": "^7.6.0",
+    "copyfiles": "^2.4.1",
     "eslint": "^8.28.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-prettier": "^8.5.0",
