+++ src/presentation/server.ts
@@ -2,6 +2,7 @@ import config from 'config';
 import { buildApp } from './index';
 import { initialize } from './logging';
 import { MdsSdk } from '@maddonkeysoftware/mds-cloud-sdk-node';
+import { ContainerManager } from '../core/container-manager';
 
 // skipcq: JS-0098
 void (async () => {
@@ -32,6 +33,14 @@ void (async () => {
     app.log.info(`Server listening at ${address}`);
 
     await Promise.all([sdkInitTask]);
+
+    // TODO: Remove this once we have a better way to start the container manager.
+    // I.e. maybe the manager should be a separate process similar to the
+    // docker minion used by serverless functions.
+    const manager =
+      app.diContainer.resolve<ContainerManager>('containerManager');
+    manager.startMonitor();
+    app.log.info('Container manager started');
   } catch (err) {
     console.error(err);
     process.exit(1);
